undo some unnecessary asynchronous hackfixing (synchronous hackfixing is enough)
