+++ packages/cli/input.json
@@ -1,7 +1,5 @@
 {
-  "tool": "inspect-data",
+  "tool": "initial-analysis",
   "params": {
-    "expression": "rules",
-    "point": "78858008544035673353062034033344525n"
   }
 }
